Make the repoURL and targetRevision propagate from the app-of-apps Application

+++ Makefile
@@ -20,15 +20,27 @@ $(LOCALBIN):
 tmp:
 	umask 0000 && mkdir -p $@
 
+##@ Deploy
+REPO_URL ?= https://github.com/kuadrant/deployment
+TARGET_REVISION ?= main
+ARGOCD_NAMESPACE ?= argocd
+deploy: yq
+	cat manifests/app-of-apps-application.yaml |\
+		$(YQ) e '.spec.source.repoURL = "$(REPO_URL)"' |\
+		$(YQ) e '.spec.source.targetRevision = "$(TARGET_REVISION)"' |\
+		$(YQ) e '.spec.destination.namespace = "$(ARGOCD_NAMESPACE)"' |\
+		kubectl -n $(ARGOCD_NAMESPACE) apply -f -
+
+
 ##@ Kind
 
 ## Targets to use kind for deployment https://kind.sigs.k8s.io
-export KUBECONFIG = $(PWD)/kubeconfig
 
 KIND_CLUSTER_NAME ?= kuadrant-local
 KIND_K8S_VERSION ?= v1.31.0@sha256:53df588e04085fd41ae12de0c3fe4c72f7013bba32a20e7325357a1ac94ba865
 .PHONY: kind-create-cluster
 kind-create-cluster-%: export KIND_EXPERIMENTAL_PROVIDER=$(CONTAINER_ENGINE)
+kind-create-cluster-%: export KUBECONFIG = $(PWD)/kubeconfig
 kind-create-cluster-%: kind kustomize ## Create the "kuadrant-local" kind cluster.
 	$(KIND) create cluster --wait 5m \
 		--image kindest/node:$(KIND_K8S_VERSION) \
@@ -37,6 +49,7 @@ kind-create-cluster-%: kind kustomize ## Create the "kuadrant-local" kind cluste
 	$(MAKE) kind-install-metallb-$*
 
 kind-install-metallb-%: export PODMAN_IGNORE_CGROUPSV1_WARNING="1"
+kind-isntall-metallb-%: export KUBECONFIG = $(PWD)/kubeconfig
 kind-install-metallb-%: yq
 	kubectl config use-context kind-kuadrant-local-$*
 	$(KUSTOMIZE) build https://github.com/metallb/metallb/config/native/?ref=v0.14.8 | kubectl apply -f -
@@ -48,31 +61,36 @@ kind-install-metallb-%: yq
 
 .PHONY: kind-delete-cluster
 kind-delete-cluster-%: export KIND_EXPERIMENTAL_PROVIDER=$(CONTAINER_ENGINE)
+kind-delete-cluster-%: export KUBECONFIG = $(PWD)/kubeconfig
 kind-delete-cluster-%: kind ## Delete the "kuadrant-local" kind cluster.
 	- $(KIND) delete cluster --name $(KIND_CLUSTER_NAME)-$*
 
+kind-apply-argocd: export KUBECONFIG = $(PWD)/kubeconfig
 kind-apply-argocd: kustomize
 	$(KUSTOMIZE) build manifests/argocd-install | yq 'select(.kind == "CustomResourceDefinition")' | kubectl apply -f -
 	sleep 2
 	kubectl wait --for condition=established --timeout=60s crd --all
 	$(KUSTOMIZE) build manifests/argocd-install | yq 'select(.kind != "CustomResourceDefinition")' | kubectl apply -f -
 
+kind-skupper-init-%: export KUBECONFIG = $(PWD)/kubeconfig
 kind-skupper-init-%: skupper
 	kubectl config use-context kind-kuadrant-local-$*
 	kubectl create namespace monitoring
 	$(SKUPPER) -n monitoring --ingress loadbalancer init
 
+kind-skupper-token-0: export KUBECONFIG = $(PWD)/kubeconfig
 kind-skupper-token-0: skupper
 	kubectl config use-context kind-kuadrant-local-0
 	$(SKUPPER) -n monitoring token create ~/skupper.token
 
+kind-skupper-link-%: export KUBECONFIG = $(PWD)/kubeconfig
 kind-skupper-link-%: skupper
 	kubectl config use-context kind-kuadrant-local-$*
 	$(SKUPPER) -n monitoring link create ~/skupper.token
 
 ##@ ArgoCD management targets
-ARGOCD_PASSWD = $(shell kubectl -n argocd get secret argocd-initial-admin-secret -o jsonpath="{.data.password}" | base64 -d)
-ARGOCD_IP = $(shell kubectl -n argocd get svc argocd-server -o jsonpath="{.status.loadBalancer.ingress[0].ip}")
+ARGOCD_PASSWD = $(shell kubectl --kubeconfig=$(PWD)/kubeconfig -n argocd get secret argocd-initial-admin-secret -o jsonpath="{.data.password}" | base64 -d)
+ARGOCD_IP = $(shell kubectl --kubeconfig=$(PWD)/kubeconfig -n argocd get svc argocd-server -o jsonpath="{.status.loadBalancer.ingress[0].ip}")
 
 argocd-url:
 	@echo -e "\n"
@@ -87,6 +105,7 @@ argocd-login: argocd
 
 ##@ Local setup
 
+local-setup: export KUBECONFIG = $(PWD)/kubeconfig
 local-setup: argocd kind-create-cluster-0 kind-create-cluster-1
 	kubectl config use-context kind-kuadrant-local-0
 	$(MAKE) kind-apply-argocd
@@ -98,6 +117,7 @@ local-setup: argocd kind-create-cluster-0 kind-create-cluster-1
                 --label deployment.kuadrant.io/hub=true && \
 		$(ARGOCD) cluster add kind-kuadrant-local-1 --name kuadrant-local-1 --yes --cluster-endpoint kube-public
 	$(MAKE) argocd-url
+	$(MAKE) deploy REPO_URL=$(REPO_URL) TARGET_REVISION=$(TARGET_REVISION) ARGOCD_NAMESPACE=$(ARGOCD_NAMESPACE)
 	$(MAKE) kind-skupper-init-0
 	$(MAKE) kind-skupper-init-1
 	$(MAKE) kind-skupper-token-0
@@ -131,8 +151,7 @@ ARGOCD_VERSION ?= v2.12.6
 ARGOCD_DOWNLOAD_URL ?= https://github.com/argoproj/argo-cd/releases/download/$(ARGOCD_VERSION)/argocd-$(OS)-$(ARCH)
 argocd: $(ARGOCD) ## Download argocd CLI locally if necessary
 $(ARGOCD): $(LOCALBIN)
-	curl -sL $(ARGOCD_DOWNLOAD_URL) -o $(ARGOCD)
-	chmod +x $(ARGOCD)
+	test -s $(ARGOCD) || curl -sL $(ARGOCD_DOWNLOAD_URL) -o $(ARGOCD) && chmod +x $(ARGOCD)
 
 ##@ Install yq
 .PHONY: yq
@@ -154,8 +173,10 @@ endif
 SKUPPER_DOWNLOAD_URL ?= https://github.com/skupperproject/skupper/releases/download/$(SKUPPER_VERSION)/skupper-cli-$(SKUPPER_VERSION)-$(OS_DL)-$(ARCH).tgz
 skupper: $(SKUPPER) ## Download skupper CLI locally if necessary
 $(SKUPPER): $(LOCALBIN)
-	curl -sL $(SKUPPER_DOWNLOAD_URL) -o $(SKUPPER).tgz
-	tar -xf $(SKUPPER).tgz -C $(LOCALBIN)
-	chmod +x $(SKUPPER)
-	touch $(SKUPPER)
-	rm $(LOCALBIN)/skupper.tgz
+	test -s $(SKUPPER) ||\
+		(curl -sL $(SKUPPER_DOWNLOAD_URL) -o $(SKUPPER).tgz &&\
+		tar -xf $(SKUPPER).tgz -C $(LOCALBIN) &&\
+		chmod +x $(SKUPPER) &&\
+		touch $(SKUPPER) &&\
+		rm $(LOCALBIN)/skupper.tgz)
+
